Refactor functions into utils module
