+++ volttron_installer/state.py
@@ -4,13 +4,13 @@ from .settings import get_settings
 from . import backend as bk
 import reflex as rx
 from .model_views import AgentModelView, ConfigStoreEntryModelView
-from .functions.create_component_uid import generate_unique_uid
+from .utils.create_component_uid import generate_unique_uid
 # from platform_page import State as PlatformPageState
 # from platform_page import Instance
 from .models import Instance
-from .functions.conversion_methods import json_string_to_csv_string, csv_string_to_json_string, identify_string_format, csv_string_to_usable_dict
-from .functions.validate_content import check_json, check_csv, check_path, check_yaml, check_regular_expression
-from .functions.create_csv_string import create_csv_string, create_and_validate_csv_string
+from .utils.conversion_methods import json_string_to_csv_string, csv_string_to_json_string, identify_string_format, csv_string_to_usable_dict
+from .utils.validate_content import check_json, check_csv, check_path, check_yaml, check_regular_expression
+from .utils.create_csv_string import create_csv_string, create_and_validate_csv_string
 import io, re, json, csv, yaml
 from .navigation.state import NavigationState
 from .backend.models import AgentType, HostEntry, PlatformConfig, PlatformDefinition, ConfigStoreEntry, AgentDefinition
@@ -19,10 +19,10 @@ from .backend.endpoints import get_all_platforms, create_platform, \
     get_agent_catalog, get_hosts, update_platform, get_inventory_service, \
     get_platform_service, get_platform_status, deploy_platform, \
     get_ansible_service
-from .functions.create_component_uid import generate_unique_uid
-from .functions.conversion_methods import csv_string_to_usable_dict
-from .functions.validate_content import check_json
-from .functions.prettify import prettify_json
+from .utils.create_component_uid import generate_unique_uid
+from .utils.conversion_methods import csv_string_to_usable_dict
+from .utils.validate_content import check_json
+from .utils.prettify import prettify_json
 from loguru import logger
 from .model_views import HostEntryModelView, PlatformModelView, AgentModelView, ConfigStoreEntryModelView, PlatformConfigModelView
 from .thin_endpoint_wrappers import *
