+++ volttron_installer/utils/prettify.py
