+++ volttron_installer/utils/validate_content.py
